Remove AutoParry component and related logic
Deleted the AutoParry MonoBehaviour and removed its instantiation and cleanup from the AutoParry patch. The code for creating and managing the parry collider is now commented out, likely in preparation for a refactor or feature removal. Also commented out the AttackToBounce patch for DamageEnemies, possibly to disable or revise its behavior.